Create a Cmake Object library for the test-suite objects.
Use this to build the test-suite and benchmark executables.
This avoids having to do any additional compilation for the benchmark.
Instead, the same objects that were used to build the test suite are
used to build the benchmark.

This also more or less obviates the need for a CMake BUILD_BENCHMARK
variable.