Relax tolerance on AmericanOption test to enable more compiler optimisations
The test tolerance is very strict and even small changes in floating point
order causes the test to fail (actual error is ~5e-15 rather than required 1e-15).
It is highly unlikely that failure in this case indicates invalid assembly.
This patch allows both AOCC with -O3 -zopt -amdveclib and ICPX with -O3 -xCORE-AVX512
-fp-model=precise to pass.